backend/requirements.txt
fastapi
uvicorn[standard]
deepface
opencv-python
python-multipart
pillow
numpy
pandas
